# openapi/fc_openapi.yaml
openapi: 3.0.1
info:
  title: Eclipse XFSC Federated Catalogue
  description: REST API of the XFSC catalogue
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://fc-server.xfsc.org
    description: some future test environment
components:
  schemas:
...
    Asset:
      type: object
      properties:
        assetHash:
          type: string
        id:
          type: string
        status:
          $ref: '#/components/schemas/AssetStatus'
        issuer:
          type: string
        validatorDids:
          description: >-
            The credentialSubjects of the parties that validated (parts of) the
            credential.
          type: array
          items:
            type: string
        uploadDatetime:
          type: string
          format: date-time
          example: '2022-03-01T13:00:00Z'
          description: The time stamp (ISO8601) when the asset was uploaded.
        statusDatetime:
          type: string
          format: date-time
          example: '2022-05-11T15:30:00Z'
          description: >-
            The last time stamp (ISO8601) the status changed (for this
            Catalogue)
        contentType:
          type: string
          description: MIME content-type of the uploaded asset (e.g. application/ld+json, application/pdf)
          example: 'application/ld+json'
        fileSize:
          type: integer
          format: int64
          description: Size of the uploaded asset in bytes
        warnings: &warningsProperty
          type: array
          description: >-
            Warnings issued during processing (e.g. triples removed due to reserved namespace filtering).
            The operation succeeded for all remaining data. Absent or empty when no warnings occurred.
          items:
            type: string
          example:
            - "2 triple(s) were removed from your upload because they use the reserved internal namespace
               <https://projects.eclipse.org/projects/technology.xfsc/federated-catalogue/meta#>.
               This namespace is reserved for internal catalogue use only and cannot be set by external clients."
        humanReadableId:
          type: string
          description: >-
            IRI of the linked human-readable representation (e.g. a PDF document),
            if one has been attached to this machine-readable asset. Absent when no link exists.
        machineReadableId:
          type: string
          description: >-
            IRI of the linked machine-readable asset, if this asset is a human-readable
            representation linked to a machine-readable parent. Absent when no link exists.
    AssetEnrichmentResponse:
      type: object
      description: Response returned when an existing non-RDF asset is enriched with RDF metadata.
      required:
        - assetId
        - triplesAdded
        - triplesRejected
      properties:
        assetId:
          type: string
          description: IRI (subjectId) of the enriched asset.
        triplesAdded:
          type: integer
          description: Number of RDF triples written to the graph store after namespace filtering.
        triplesRejected:
          type: integer
          description: Number of triples filtered out due to protected namespace policy.
    AssetResult:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Asset'
        content:
          type: string
    Assets:
      allOf:
        - $ref: '#/components/schemas/ResultPage'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AssetResult'
    AssetStatus:
      type: string
      enum:
        - active
        - eol
        - deprecated
        - revoked
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    ResultPage:
      type: object
      properties:
        totalCount:
          type: integer
...
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >-
        Forbidden. The user does not have the permission to execute this
        request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: >-
        May contain hints how to solve the error or indicate what went wrong at
        the server. Must not outline any information about the internal
        structure of the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ClientError:
      description: >-
        May contain hints how to solve the error or indicate what was wrong in
        the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: HTTP Conflict 409
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  parameters:
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      required: false
      description: The number of items to skip before starting to collect the result set.
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      required: false
      description: The number of items to return.
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /assets:
    get:
      tags:
        - Assets
      summary: Get the list of asset metadata in the Catalogue
      operationId: readAssets
      security:
        - jwt: []
      parameters:
        - name: uploadTimerange
          in: query
          description: >-
            Filter for the time range when the asset was uploaded to the
            catalogue. The time range has to be specified as start time and
            end time as ISO8601 timestamp separated by a `/`.
          required: false
          example: 2022-03-01T13:00:00Z/2022-05-11T15:30:00Z
          schema:
            type: string
        - name: statusTimerange
          in: query
          description: >-
            Filter for the time range when the status of the asset was last
            changed in the catalogue. The time range has to be specified as
            start time and end time as ISO8601 timestamp separated by a `/`.
          required: false
          example: 2022-03-01T13:00:00Z/2022-05-11T15:30:00Z
          schema:
            type: string
        - name: issuers
          in: query
          description: >-
            Filter for the issuer of the credential. This is the unique ID of the
            Participant that has prepared the credential.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false
        - name: validators
          in: query
          description: >-
            Filter for a validator of the credential. This is the unique ID of the
            Participant that validated (part of) the credential.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false
        - name: statuses
          in: query
          description: Filter for the status of the asset.
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/AssetStatus'
          explode: false
        - name: ids
          in: query
          description: Filter for the id/credentialSubject of the credential.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false
        - name: hashes
          in: query
          description: Filter for the hash of the asset.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false
        - name: withMeta
          in: query
          description: Whether to include asset metadata in the response.
          required: false
          schema:
            type: boolean
            default: true
        - name: withContent
          in: query
          description: Whether to include asset content in the response.
          required: false
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of metadata of available assets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assets'
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - Assets
      summary: Add a new asset to the catalogue
      description: >-
        Upload an asset. Non-RDF content is stored directly (HTTP 201).
        RDF content is automatically detected and routed: if the RDF subject IRI
        matches an existing non-RDF asset, the asset is enriched with metadata
        (HTTP 200, returns AssetEnrichmentResponse); otherwise the RDF is
        stored as a new or updated credential asset (HTTP 201, returns Asset).
        Human-readable assets (attachments uploaded via
        `/assets/{asset_id}/human-readable`) cannot be enriched with metadata —
        attempting to enrich one returns HTTP 422.
        Trust-framework base-class compliance is not enforced on upload: a
        credential whose subject type does not resolve to a known base class is
        still stored (HTTP 201). Use `POST /verification` with
        `requireBaseClass=true` to enforce base-class resolution.
      operationId: addAsset
      security:
        - jwt: []
      requestBody:
        description: >-
          The asset content. For JSON-LD credentials, use application/json,
          application/vc+ld+json, or application/vp+ld+json. For JWT-wrapped
          credentials use application/vc+jwt (VC) or application/vp+jwt (VP) —
          these are IANA-registered types per W3C VC-JOSE-COSE and are issued by
          the live GXDCH Compliance Service. The ICAM 24.07 variants
          application/vc+ld+json+jwt and application/vp+ld+json+jwt are also
          accepted for backward compatibility but are not IANA-registered.
          For non-credential RDF serializations, use text/turtle,
          application/n-triples, or application/rdf+xml — triples are extracted
          and stored in the graph, but schema validation is not performed.
          For non-RDF assets, use multipart/form-data or application/octet-stream
          (handled by the server outside the generated delegate pattern).
        content:
          application/json:
            schema:
              type: string
          application/ld+json:
            schema:
              type: string
          application/vc+ld+json:
            schema:
              type: string
          application/vp+ld+json:
            schema:
              type: string
          application/vc+jwt:
            schema:
              type: string
          application/vp+jwt:
            schema:
              type: string
          application/vc+ld+json+jwt:
            schema:
              type: string
          application/vp+ld+json+jwt:
            schema:
              type: string
          text/turtle:
            schema:
              type: string
          application/n-triples:
            schema:
              type: string
          application/rdf+xml:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: >-
            Non-RDF asset enriched with RDF metadata. Returned when the uploaded
            RDF payload matches the subject IRI of an existing non-RDF asset.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetEnrichmentResponse'
        '201':
          description: Created
          headers:
            location:
              description: URL of the added asset.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          description: >-
            Unprocessable entity. Returned when the upload describes a subject IRI
            that resolves to a human-readable attachment, which cannot be enriched
            with metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /assets/{id}:
    get:
      tags:
        - Assets
      summary: Read an asset by its IRI.
      description: >
        Returns a JSON-serialized Asset metadata object for any asset identified by its IRI.

        For RDF (JSON-LD) assets, the rawContent field is populated with the raw JSON-LD string.

        For any asset (RDF or non-RDF) that has bidirectional links to another representation,
        the humanReadableId or machineReadableId fields are populated.

        For non-RDF assets (PDF, images, binary files), the file content can be retrieved via
        GET /assets/{id}/human-readable or GET /assets/{id}/machine-readable.
      operationId: readAssetById
      parameters:
        - name: id
          in: path
          description: IRI identifier of the asset (DID, URN, or HTTP IRI)
          required: true
          schema:
            type: string
          examples:
            did:
              summary: DID Web identifier
              value: 'did:web:example.com:participant'
            urn:
              summary: UUID URN (non-RDF assets)
              value: 'urn:uuid:550e8400-e29b-41d4-a716-446655440000'
            http:
              summary: HTTP IRI
              value: 'https://example.org/asset/123'
        - name: version
          in: query
          description: Optional 1-based version ordinal. If provided, returns the asset content at that specific version.
          required: false
          schema:
            type: integer
            minimum: 1
      security:
        - jwt: []
      responses:
        '200':
          description: >
            The asset metadata as a JSON-serialized Asset object.
            For RDF (JSON-LD) assets, the rawContent field is populated with the raw JSON-LD string.
            For any asset with bidirectional links, humanReadableId or machineReadableId fields are populated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - Assets
      summary: Update an asset in-place (versioned).
      description: >
        Updates an existing asset identified by its IRI. The update is applied in-place;
        Envers records the previous state as a new revision in assets_aud automatically.
        If no asset exists for the given IRI, the asset is created as version 1.
        Required permission: ASSET_UPDATE or ADMIN_ALL
      operationId: updateAsset
      parameters:
        - name: id
          in: path
          description: IRI identifier of the asset
          required: true
          schema:
            type: string
        - name: changeComment
          in: query
          description: Optional human-readable note for this version.
          required: false
          schema:
            type: string
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/ld+json:
            schema:
              type: string
          application/json:
            schema:
              type: string
          text/turtle:
            schema:
              type: string
          application/n-triples:
            schema:
              type: string
          application/rdf+xml:
            schema:
              type: string
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /assets/{asset_hash}:
    delete:
      tags:
        - Assets
      summary: Completely delete an asset by its content hash
      description: >
        Unlike other asset endpoints which use IRI-based identification (/assets/{id}),
        the delete operation uses the SHA-256 content hash to guarantee unambiguous
        single-row targeting. See ADR 7 — Delete Operation Exception.


        For idempotent cascade-by-IRI semantics (every version, every linked companion, every
        provenance row), prefer DELETE /assets/by-id/{id}.
      operationId: deleteAsset
      security:
        - jwt: [ ]
      parameters:
        - name: asset_hash
          in: path
          description: Hash identifier of the asset
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...